refactor: rename "ws" plugin to "legacyWebsocket"
