+++ packages/broker/configs/development-1.env.json
@@ -33,7 +33,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": {
+    "legacyWebsocket": {
       "port": 8890,
       "pingInterval": 60000
     },
