+++ packages/broker/configs/development-2.env.json
@@ -33,7 +33,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": { 
+    "legacyWebsocket": { 
       "port": 8790
     },
     "legacyPublishHttp": {},
