+++ packages/broker/configs/development-3.env.json
@@ -33,7 +33,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": {
+    "legacyWebsocket": {
       "port": 8690
     },
     "legacyPublishHttp": {},
