+++ packages/broker/configs/docker-1.env.json
@@ -29,7 +29,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": {
+    "legacyWebsocket": {
       "port": 8890
     },
     "legacyPublishHttp": {},
