+++ packages/broker/configs/docker-2.env.json
@@ -32,7 +32,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": {
+    "legacyWebsocket": {
       "port": 8790
     },
     "legacyPublishHttp": {},
