+++ packages/broker/configs/docker-3.env.json
@@ -32,7 +32,7 @@
   },
   "apiAuthentication": null,
   "plugins": {
-    "ws": {
+    "legacyWebsocket": {
       "port": 8690
     },
     "legacyPublishHttp": {},
