+++ packages/broker/src/localStreamrClient.ts
@@ -5,10 +5,10 @@ import { Config } from './config'
 // This is a temporary solution to prototype Brubeck-plugins
 
 export const createLocalStreamrClient = (config: Config): StreamrClient|undefined => {
-    if (!config.plugins.ws) {
+    if (!config.plugins.legacyWebsocket) {
         return undefined
     }
-    const wsPort = config.plugins.ws.port
+    const wsPort = config.plugins.legacyWebsocket.port
     return new StreamrClient({
         auth:{
             privateKey: config.ethereumPrivateKey
