+++ packages/broker/src/pluginRegistry.ts
@@ -2,7 +2,7 @@ import { PluginOptions } from './Plugin'
 import { PublishHttpPlugin } from './plugins/publishHttp/PublishHttpPlugin'
 import { PublishHttpPlugin as LegacyPublishHttpPlugin } from './plugins/legacyPublishHttp/PublishHttpPlugin'
 import { MetricsPlugin } from './plugins/metrics/MetricsPlugin'
-import { WebsocketPlugin } from './plugins/websocket/WebsocketPlugin'
+import { WebsocketPlugin as LegacyWebsocketPlugin } from './plugins/legacyWebsocket/WebsocketPlugin'
 import { MqttPlugin } from './plugins/mqtt/MqttPlugin'
 import { MqttPlugin as LegacyMqttPlugin } from './plugins/legacyMqtt/MqttPlugin'
 import { StoragePlugin } from './plugins/storage/StoragePlugin'
@@ -15,8 +15,8 @@ export const createPlugin = (name: string, pluginOptions: PluginOptions) => {
             return new LegacyPublishHttpPlugin(pluginOptions)
         case 'metrics':
             return new MetricsPlugin(pluginOptions)
-        case 'ws':
-            return new WebsocketPlugin(pluginOptions)
+        case 'legacyWebsocket':
+            return new LegacyWebsocketPlugin(pluginOptions)
         case 'mqtt':
             return new MqttPlugin(pluginOptions)
         case 'legacyMqtt':
