+++ packages/broker/src/plugins/legacyWebsocket/config.schema.json
