+++ packages/broker/test/integration/plugins/storage/resends-cancelled-on-client-disconnect.test.ts
@@ -8,7 +8,7 @@ import { once } from 'events'
 import { Wallet } from 'ethers'
 import { wait } from 'streamr-test-utils'
 import { PassThrough } from 'stream'
-import { WebsocketServer } from '../../../../src/plugins/websocket/WebsocketServer'
+import { WebsocketServer } from '../../../../src/plugins/legacyWebsocket/WebsocketServer'
 import { StreamFetcher } from '../../../../src/StreamFetcher'
 import { Publisher } from '../../../../src/Publisher'
 import { SubscriptionManager } from '../../../../src/SubscriptionManager'
